CA2_devOps/requirements.txt
scikit-learn == 1.0.2
Flask == 2.2.2
Flask-SQLAlchemy == 3.0.2
WTForms == 3.0.1
Flask-WTF == 1.0.1
Flask-Login == 0.5.0
jinja2 == 3.1.2
SQLAlchemy == 1.4.44
wcwidth ==  0.2.5
Werkzeug == 2.2.2
pickleshare == 0.7.5
pickle-mixin == 1.0.2
pandas 
pytest == 7.1.2
gunicorn == 20.1.0
tensorflow 
Image
Pillow
numpy
requests